Revamp the setup-environment protocol for exchanging environment information
The current setup-environment architecture assumes two internal
programs: setup-environment-internal.sh (s-i-e.sh) and
setup-environment-internal.py (s-i-e.py).  s-e-i.sh calls s-e-i.py.

In the previous scheme, s-e-i.py would create a temporary file to
store the environment and report its path back to s-i-e.sh as the last
line printed to the standard output.  That meant that s-i-e.sh would
capture all the s-i-e.py standard output in a variable, just to find
out what file the environment has been printed to.  That is rather
awkward, since it cripples s-e-i.py in the sense that it cannot
normally use the standard output.

In the new scheme, implemented by this commit, s-e-i.sh creates the
temporary file to hold the environment and give s-e-i.py it.  With
this approach, s-e-i.py can freely use the standard output.  The
environment will be printed to the file given by s-e-i.sh.  The
standard output is no longer the "transport" for the environment
information exchange -- it's all done via command line arguments and a
temporary file.

Change-Id: I8d45c691e27914e1213614988e70edb78b2ec85d
Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>