Read a `priority' file for projects that are not proper layers
This commit implements an alternative way of specifying project
priority for projects that are not proper Yocto Project layers.

Projects like yocto-config and ye are not proper Yocto Project layers,
so they have no BBFILE_PRIORITY_<layer> configuration.  They can,
however, provide hook scripts.  setup-environment requires a
per-project priority information to be able to run hook scripts in
deterministic order.  Only a single project is allowed to have no
priority; in this case, it is assumed to have the highest priority.

Priorities for projects that are not proper Yocto Project layers can
be specified by a `priority' file in the projects' setup-environment.d
directory.  The `priority' file must contain an integer indicating the
project priority.

Example:

$ basename `pwd`
setup-environment.d

$ ls -l
total 12
-rw-rw-r-- 1 mario ossbot   1 Dec  9 10:46 priority
-rw-r--r-- 1 mario ossbot 243 Dec  9 10:38 ye.py

$ cat priority
7

Change-Id: I700de2863dc00d7b59db8c4bf52326df793eb6a8
Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>