Update patch set 3
Patch Set 3:

I'm not completely sure, but as far as I can see, `count_trailing_spaces' would break when given an empty string.  A simple fix would be testing if the given argument is an empty string and return 0 in that case.

Reversing the string then using `count_leading_spaces' works too, although is less efficient (but that probably is not really relevant in this case).

Patch-set: 3